*Removing default properties to not debug zuul; *Removing not necessary class; *Fix trace filter order
